# docs/swagger.yaml
basePath: /
...
host: 127.0.0.1:11470
info:
  contact: {}
  description: HTTP API served by stremio-server-go, a pure-Go drop-in for Stremio's
    streaming server (server.js). serverVersion is reported as 4.21.0 for client feature-gating;
    it is independent of the binary build version.
  license:
    name: MIT
    url: https://github.com/M0Rf30/stremio-server-go/blob/main/LICENSE
  title: stremio-server-go enginefs API
  version: 4.21.0
paths:
...
  /casting/{id}/player:
    get:
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      - description: load|play|pause|stop|seek|status
        in: query
        name: command
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
      summary: UPnP AVTransport control
      tags:
      - Casting
    post:
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      - description: load|play|pause|stop|seek|status
        in: query
        name: command
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
      summary: UPnP AVTransport control
      tags:
      - Casting
...
schemes:
- http
- https
swagger: "2.0"